backend/requirements.txt
# Backend Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Cache
redis>=5.0.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Utils
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.27.0
aiofiles>=24.1.0

# MinIO S3
minio>=7.2.0

# Vector Database (future)
weaviate-client>=4.5.0
holidays>=0.40.0